requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
annotated-types==0.6.0
anykeystore==0.2
apex==0.9.10.dev0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cryptacular==1.6.2
cycler==0.12.1
datasets==2.21.0
defusedxml==0.7.1
dill==0.3.6
docstring_parser==0.16
einops==0.8.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.2
filelock==3.9.0
flash-attn==2.6.3
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.11.0
huggingface-hub==0.24.5
hupper==1.12.1
idna==3.6
imageio==2.34.1
importlib_metadata==7.1.0
inflate64==1.0.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.4.2
kaleido==0.2.1
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning==2.2.4
lightning-utilities==0.11.2
littleutils==0.2.2
loralib==0.1.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.3.2
mlxtend==0.23.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.14
multivolumefile==0.2.3
namex==0.0.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outdated==0.2.2
packaging==24.0
pandas==2.2.1
PasteDeploy==3.1.0
pbkdf2==1.3
peft==0.12.0
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
plotly==5.22.0
portalocker==2.8.2
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
ptflops==0.7.3
py7zr==0.22.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybcj==1.0.2
pycocotools==2.0.7
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
pyparsing==3.1.2
pyppmd==1.1.0
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.1
pyzstd==0.16.1
regex==2023.12.25
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
rich==13.7.1
rouge_score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
shtab==1.7.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
sympy==1.12
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
texttable==1.7.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0+cu121
torch-fidelity==0.3.0
torch_geometric==2.5.3
torchaudio==2.4.0+cu121
torchmetrics==1.4.0.post0
torchsummary==1.5.1
torchvision==0.19.0+cu121
tqdm==4.66.5
transaction==4.0
transformers==4.44.0
translationstring==1.4
triton==3.0.0
trl==0.9.6
typeguard==2.13.3
typing_extensions==4.12.2
tyro==0.8.5
tzdata==2024.1
urllib3==2.2.1
velruse==1.1.1
venusian==3.1.0
wcwidth==0.2.13
WebOb==1.8.8
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zope.deprecation==5.0
zope.interface==7.0.1
zope.sqlalchemy==3.1
